+++ src/peek.ts
@@ -340,7 +340,7 @@ type DialogProperties = Record<string, string>;
         _configuration.copyText = Data.getDefaultAnyString( _configuration.copyText, "Copy" );
         _configuration.copySymbolText = Data.getDefaultAnyString( _configuration.copySymbolText, "❐" );
         _configuration.pasteText = Data.getDefaultAnyString( _configuration.pasteText, "Paste" );
-        _configuration.pasteSymbolText = Data.getDefaultAnyString( _configuration.pasteSymbolText, "+" );
+        _configuration.pasteSymbolText = Data.getDefaultAnyString( _configuration.pasteSymbolText, "☐" );
         _configuration.removeText = Data.getDefaultAnyString( _configuration.removeText, "Remove" );
         _configuration.removeSymbolText = Data.getDefaultAnyString( _configuration.removeSymbolText, "✕" );
     }