+++ docs/CHANGE_LOG.md
@@ -10,9 +10,9 @@
 - Added configuration option "copyText" (defaults to "Copy"), which states the text to use for the "Copy" button.
 - Added configuration option "copySymbolText" (defaults to "❐"), which states the symbol text to use for the "Copy" button.
 - Added configuration option "pasteText" (defaults to "Paste"), which states the text to use for the "Paste" button.
-- Added configuration option "pasteSymbolText" (defaults to "+"), which states the symbol text to use for the "Paste" button.
+- Added configuration option "pasteSymbolText" (defaults to "☐"), which states the symbol text to use for the "Paste" button.
 - Added configuration option "removeText" (defaults to "Remove"), which states the text to use for the "Remove" button.
-- Added configuration option "removeSymbolText" (defaults to "✖"), which states the symbol text to use for the "Remove" button.
+- Added configuration option "removeSymbolText" (defaults to "✕"), which states the symbol text to use for the "Remove" button.
 
 #### **UI Improvements:**
 - The property name label now uses a ellipsis effect when the property name is long.