+++ docs/CHANGE_LOG.md
@@ -16,6 +16,7 @@
 
 #### **UI Improvements:**
 - The property name label now uses a ellipsis effect when the property name is long.
+- Added a 2000 z-index to the dialog, to make sure its on top of most elements.
 
 <br>